Speed up YARP.lex_compat
