+++ bin/lex
@@ -12,8 +12,8 @@ puts pattern % ["Ripper lex", "YARP lex"]
 puts pattern % ["-" * 70, "-" * 70]
 
 source = File.read(filepath)
-ripper = YARP.remove_tilde_heredocs(YARP.lex_ripper(source))
-yarp = YARP.remove_tilde_heredocs(YARP.lex_compat(source).value)
+ripper = YARP.lex_ripper(source)
+yarp = YARP.lex_compat(source).value
 
 [yarp.length, ripper.length].max.times do |index|
   left = ripper[index]
