+++ lib/yarp.rb
@@ -1,7 +1,5 @@
 # frozen_string_literal: true
 
-require "delegate"
-
 module YARP
   # This represents a location in the source corresponding to a node or token.
   class Location
@@ -167,398 +165,7 @@ module YARP
   # This lexes with the Ripper lex. It drops any space events but otherwise
   # returns the same tokens.
   def self.lex_ripper(source)
-    Ripper.lex(source).each_with_object([]) do |token_arr, tokens|
-      token = LexedToken.new(token_arr)
-      tokens << token unless token.event == :on_sp
-    end
-  end
-
-  class LexedToken < SimpleDelegator
-    def location
-      self[0]
-    end
-
-    def event
-      self[1]
-    end
-
-    def value
-      self[2]
-    end
-
-    def state
-      self[3]
-    end
-
-    def state=(val)
-      self[3] = val
-    end
-
-    def ==(other)
-      if event == :on_comment
-        location == other.location && event == other.event && value == other.value
-      elsif event == :on_ident
-        left, right = [self[3], other[3]].map(&:to_i).sort
-
-        # In the event that we're comparing identifiers, we're going to allow a
-        # little divergence. Ripper doesn't account for local variables
-        # introduced through named captures in regexes.
-        if [Ripper::EXPR_ARG, Ripper::EXPR_CMDARG].include?(left) && right == Ripper::EXPR_END | Ripper::EXPR_LABEL
-          self[0...-1] == other[0...-1]
-        else
-          super
-        end
-      else
-        super
-      end
-    end
-
-    def is_tilde_heredoc?
-      value[2] == "~"
-    end
-  end
-
-  # This class is responsible for lexing the source using YARP and then
-  # converting those tokens to be compatible with Ripper.
-  class LexCompat
-    RIPPER = {
-      AMPERSAND: :on_op,
-      AMPERSAND_AMPERSAND: :on_op,
-      AMPERSAND_AMPERSAND_EQUAL: :on_op,
-      AMPERSAND_DOT: :on_op,
-      AMPERSAND_EQUAL: :on_op,
-      BACK_REFERENCE: :on_backref,
-      BACKTICK: :on_backtick,
-      BANG: :on_op,
-      BANG_EQUAL: :on_op,
-      BANG_TILDE: :on_op,
-      BRACE_LEFT: :on_lbrace,
-      BRACE_RIGHT: :on_rbrace,
-      BRACKET_LEFT: :on_lbracket,
-      BRACKET_LEFT_ARRAY: :on_lbracket,
-      BRACKET_LEFT_RIGHT: :on_op,
-      BRACKET_LEFT_RIGHT_EQUAL: :on_op,
-      BRACKET_RIGHT: :on_rbracket,
-      CARET: :on_op,
-      CARET_EQUAL: :on_op,
-      CHARACTER_LITERAL: :on_CHAR,
-      CLASS_VARIABLE: :on_cvar,
-      COLON: :on_op,
-      COLON_COLON: :on_op,
-      COMMA: :on_comma,
-      COMMENT: :on_comment,
-      CONSTANT: :on_const,
-      DOT: :on_period,
-      DOT_DOT: :on_op,
-      DOT_DOT_DOT: :on_op,
-      EMBDOC_BEGIN: :on_embdoc_beg,
-      EMBDOC_END: :on_embdoc_end,
-      EMBDOC_LINE: :on_embdoc,
-      EMBEXPR_BEGIN: :on_embexpr_beg,
-      EMBEXPR_END: :on_embexpr_end,
-      EMBVAR: :on_embvar,
-      EOF: :on_eof,
-      EQUAL: :on_op,
-      EQUAL_EQUAL: :on_op,
-      EQUAL_EQUAL_EQUAL: :on_op,
-      EQUAL_GREATER: :on_op,
-      EQUAL_TILDE: :on_op,
-      FLOAT: :on_float,
-      GREATER: :on_op,
-      GREATER_EQUAL: :on_op,
-      GREATER_GREATER: :on_op,
-      GREATER_GREATER_EQUAL: :on_op,
-      GLOBAL_VARIABLE: :on_gvar,
-      HEREDOC_END: :on_heredoc_end,
-      HEREDOC_START: :on_heredoc_beg,
-      IDENTIFIER: :on_ident,
-      IGNORED_NEWLINE: :on_ignored_nl,
-      IMAGINARY_NUMBER: :on_imaginary,
-      INTEGER: :on_int,
-      INSTANCE_VARIABLE: :on_ivar,
-      INVALID: :INVALID,
-      KEYWORD___ENCODING__: :on_kw,
-      KEYWORD___LINE__: :on_kw,
-      KEYWORD___FILE__: :on_kw,
-      KEYWORD_ALIAS: :on_kw,
-      KEYWORD_AND: :on_kw,
-      KEYWORD_BEGIN: :on_kw,
-      KEYWORD_BEGIN_UPCASE: :on_kw,
-      KEYWORD_BREAK: :on_kw,
-      KEYWORD_CASE: :on_kw,
-      KEYWORD_CLASS: :on_kw,
-      KEYWORD_DEF: :on_kw,
-      KEYWORD_DEFINED: :on_kw,
-      KEYWORD_DO: :on_kw,
-      KEYWORD_DO_LOOP: :on_kw,
-      KEYWORD_ELSE: :on_kw,
-      KEYWORD_ELSIF: :on_kw,
-      KEYWORD_END: :on_kw,
-      KEYWORD_END_UPCASE: :on_kw,
-      KEYWORD_ENSURE: :on_kw,
-      KEYWORD_FALSE: :on_kw,
-      KEYWORD_FOR: :on_kw,
-      KEYWORD_IF: :on_kw,
-      KEYWORD_IN: :on_kw,
-      KEYWORD_MODULE: :on_kw,
-      KEYWORD_NEXT: :on_kw,
-      KEYWORD_NIL: :on_kw,
-      KEYWORD_NOT: :on_kw,
-      KEYWORD_OR: :on_kw,
-      KEYWORD_REDO: :on_kw,
-      KEYWORD_RESCUE: :on_kw,
-      KEYWORD_RETRY: :on_kw,
-      KEYWORD_RETURN: :on_kw,
-      KEYWORD_SELF: :on_kw,
-      KEYWORD_SUPER: :on_kw,
-      KEYWORD_THEN: :on_kw,
-      KEYWORD_TRUE: :on_kw,
-      KEYWORD_UNDEF: :on_kw,
-      KEYWORD_UNLESS: :on_kw,
-      KEYWORD_UNTIL: :on_kw,
-      KEYWORD_WHEN: :on_kw,
-      KEYWORD_WHILE: :on_kw,
-      KEYWORD_YIELD: :on_kw,
-      LABEL: :on_label,
-      LAMBDA_BEGIN: :on_tlambeg,
-      LESS: :on_op,
-      LESS_EQUAL: :on_op,
-      LESS_EQUAL_GREATER: :on_op,
-      LESS_LESS: :on_op,
-      LESS_LESS_EQUAL: :on_op,
-      MINUS: :on_op,
-      MINUS_EQUAL: :on_op,
-      MINUS_GREATER: :on_tlambda,
-      NEWLINE: :on_nl,
-      NTH_REFERENCE: :on_backref,
-      PARENTHESIS_LEFT: :on_lparen,
-      PARENTHESIS_RIGHT: :on_rparen,
-      PERCENT: :on_op,
-      PERCENT_EQUAL: :on_op,
-      PERCENT_LOWER_I: :on_qsymbols_beg,
-      PERCENT_LOWER_W: :on_qwords_beg,
-      PERCENT_LOWER_X: :on_backtick,
-      PERCENT_UPPER_I: :on_symbols_beg,
-      PERCENT_UPPER_W: :on_words_beg,
-      PIPE: :on_op,
-      PIPE_EQUAL: :on_op,
-      PIPE_PIPE: :on_op,
-      PIPE_PIPE_EQUAL: :on_op,
-      PLUS: :on_op,
-      PLUS_EQUAL: :on_op,
-      QUESTION_MARK: :on_op,
-      RATIONAL_NUMBER: :on_rational,
-      REGEXP_BEGIN: :on_regexp_beg,
-      REGEXP_END: :on_regexp_end,
-      SEMICOLON: :on_semicolon,
-      SLASH: :on_op,
-      SLASH_EQUAL: :on_op,
-      STAR: :on_op,
-      STAR_EQUAL: :on_op,
-      STAR_STAR: :on_op,
-      STAR_STAR_EQUAL: :on_op,
-      STRING_BEGIN: :on_tstring_beg,
-      STRING_CONTENT: :on_tstring_content,
-      STRING_END: :on_tstring_end,
-      SYMBOL_BEGIN: :on_symbeg,
-      TILDE: :on_op,
-      UMINUS: :on_op,
-      UPLUS: :on_op,
-      WORDS_SEP: :on_words_sep,
-      __END__: :on___end__
-    }.freeze
-
-    # A heredoc in this case is a list of tokens that belong to the body of the
-    # heredoc that should be appended onto the list of tokens when the heredoc
-    # closes.
-    class Heredoc
-      attr_reader :state, :tokens
-
-      def initialize(state)
-        @state = state
-        @tokens = []
-      end
-
-      def <<(token)
-        tokens << token
-      end
-
-      def to_a
-        tokens.last.state = state
-        tokens
-      end
-    end
-
-    private_constant :RIPPER, :Heredoc
-
-    attr_reader :source, :offsets
-
-    def initialize(source)
-      @source = source
-
-      @offsets = [0]
-      last_offset = 0
-
-      source.each_line do |line|
-        last_offset += line.bytesize
-        @offsets << last_offset
-      end
-    end
-
-    def result
-      tokens = []
-
-      state = :default
-      heredocs = []
-
-      result = YARP.lex(source)
-      previous_state = nil
-
-      # If there's a UTF-8 byte-order mark as the start of the file, then ripper
-      # sets every token's on the first line back by 6 bytes. It also keeps the
-      # byte order mark in the first token's value. This is weird, and I don't
-      # want to mirror that in our parser. So instead, we'll match up the values
-      # here, and then match up the locations as we process the tokens.
-      bom = source.bytes[0..2] == [0xEF, 0xBB, 0xBF]
-      result.value[0][0].value.prepend("\xEF\xBB\xBF") if bom
-
-      result.value.each_with_index do |(token, lex_state), index|
-        event = RIPPER.fetch(token.type)
-        lex_state =
-          # On regex end, Ripper scans and then sets end state, so the ripper
-          # lexed output is begin, when it should be end. YARP sets lex state
-          # correctly to end state, but we want to be able to compare against
-          # Ripper's lexed state. So here, if it's a regexp end token, we output
-          # the state as the previous state, solely for the sake of comparison.
-          if event == :on_regexp_end
-            previous_token = result.value[index - 1][0]
-
-            if RIPPER.fetch(previous_token.type) == :on_embexpr_end
-              # If the previous token is embexpr_end, then we have to do even
-              # more processing. The end of an embedded expression sets the
-              # state to the state that it had at the beginning of the embedded
-              # expression. So we have to go and find that state and set it
-              # here.
-              counter = 1
-              current_index = index - 1
-
-              until counter == 0
-                current_index -= 1
-                current_event = RIPPER.fetch(result.value[current_index][0].type)
-                counter += { on_embexpr_beg: -1, on_embexpr_end: 1 }[current_event] || 0
-              end
-
-              Ripper::Lexer::State.new(result.value[current_index][1])
-            else
-              previous_state
-            end
-          else
-            Ripper::Lexer::State.new(lex_state)
-          end
-
-        (lineno, column) = location_for(token.location.start_offset)
-        column -= 6 if bom && lineno == 1
-
-        token = LexedToken.new(
-          [
-            [lineno, column],
-            event,
-            value_for(event, token.value),
-            lex_state
-          ]
-        )
-
-        previous_state = lex_state
-
-        # The order in which tokens appear in our lexer is different from the
-        # order that they appear in Ripper. When we hit the declaration of a
-        # heredoc in YARP, we skip forward and lex the rest of the content of
-        # the heredoc before going back and lexing at the end of the heredoc
-        # identifier.
-        #
-        # To match up to ripper, we keep a small state variable around here to
-        # track whether we're in the middle of a heredoc or not. In this way we
-        # can shuffle around the token to match Ripper's output.
-        case state
-        when :default
-          tokens << token
-
-          if event == :on_heredoc_beg
-            state = :heredoc_opened
-            heredocs << Heredoc.new(lex_state)
-          end
-        when :heredoc_opened
-          heredocs.last << token
-          state = :heredoc_closed if event == :on_heredoc_end
-        when :heredoc_closed
-          tokens << token
-
-          case event
-          when :on_nl, :on_ignored_nl
-            heredocs.each do |heredoc|
-              tokens.concat(heredoc.to_a)
-            end
-
-            heredocs.clear
-            state = :default
-          when :on_heredoc_beg
-            state = :heredoc_opened
-            heredocs << Heredoc.new(lex_state)
-          end
-        end
-      end
-
-      ParseResult.new(tokens[0...-1], result.comments, result.errors, result.warnings)
-    end
-
-    private
-
-    def location_for(value)
-      line_number = offsets.bsearch_index { |offset| offset > value }
-      line_offset = offsets[line_number - 1] if line_number
-
-      [
-        line_number || offsets.length - 1,
-        value - (line_offset || offsets.last)
-      ]
-    end
-
-    def value_for(event, value)
-      if event == :on___end__
-        # Ripper doesn't include the rest of the token in the event, so we
-        # need to trim it down to just the first newline.
-        value[0..value.index("\n")]
-      else
-        value
-      end
-    end
-  end
-
-  # Returns an array of tokens that closely resembles that of the Ripper lexer.
-  # The only difference is that since we don't keep track of lexer state in the
-  # same way, it's going to always return the NONE state.
-  def self.lex_compat(source)
-    LexCompat.new(source).result
-  end
-
-  # We handle tilde heredocs in the parsing phase, not the lexing phase
-  # However, we want to preserve the usefulness of comparing lexed results
-  # This method allows us to sanitize lexing so we can ignore any differences
-  # in the tilde heredocs
-  def self.remove_tilde_heredocs(tokens)
-    res = []
-
-    ignoring_tokens = false
-    tokens.each do |token|
-      if token.event == :on_heredoc_beg && token.is_tilde_heredoc?
-        ignoring_tokens = true
-      elsif token.event == :on_heredoc_end
-        ignoring_tokens = false
-      elsif !ignoring_tokens
-        res << token
-      end
-    end
-    res
+    Ripper.lex(source).reject { |token| token[1] == :on_sp }
   end
 
   # Load the serialized AST using the source as a reference into a tree.
@@ -567,6 +174,7 @@ module YARP
   end
 end
 
+require_relative "yarp/lex_compat"
 require_relative "yarp/node"
 require_relative "yarp/ripper_compat"
 require_relative "yarp/serialize"
